Remove IVariableInfo dependecy from ProcessedType, overall still WIP
* Removed IVariableInfo dictionary from ProcessedType. It created a circular
reference between immutable types. The variable for a field is not a property of
the ProcessedType either, that's the MemoryManager's job to handle.

* Organized built-in types in CompilationState so they're easy to access.
Removed getting ProcessedTypes by string.

* Added MemoryMap class for tracking where variables are laid out in memory.
MemoryManager returns one instead of putting the IVariableInfo in ProcessedType.

* Literals are evaluated without reflection now.

* TypeParser no longer uses Type or TypeInfo. TypeInfo is still used for
checking ReservedSymbols' fields, but that's appropriate.

* Static field dictionary in ProcessedType replaced with an array. No longer
need to pass field info into the constructor.

- Output still correct. Unit tests still broken.